PDD/PDD-PZ2/requirements.txt
pyspark
PyDrive
numpy
pandas
scipy
jupyter
kafka-python
pyarrow
